Separated the event listeners into folders
Relates #9
